Start a PEAR package with utility code to access package lists from PEAR servers.
The package does not intend to be a replacement for the PEAR libraries. It will only provide tools to deal with simple functionality that can be coded quickly. The intention is to have a cleaner development API to deal with PEAR related stuff. Neither PEAR nor Pyrus (PEAR2) made me happy so far.